BREAKING: Move ExtensionError into hugr-core/src/hugr/
